requirements.txt
langchain
streamlit
langsmith
python-dotenv
aiogram
